Added test when launching with existing valid login
